pager: expose QueryPager::next() under flag
If `unstable-csharp-rs` feature is enabled, then
`QueryPager::next_column_iterator()` is available and public, and is
essentially a wrapper around `QueryPager::next()`.

Rationale: `QueryPager::next()` is the only way to achieve borrowed
deserialization using `QueryPager`, but it is private.
C# wrapper will benefit from having access to it, so I made it
available under a dedicated feature, similarly to how `unstable-cpp-rs`
feature works.